Set Functions and New Fail State
Turned lists into set lists, for built-in set operator over-rides. Added
in an error state for non-existing elements. Fails quietly most of the
time, because I haven't come up with a good idea of how to cancel out
the error/warning/stdout.write messages that popup in Maya's global
readout.